+++ config/environment.rb
@@ -1,10 +1,8 @@
 # frozen_string_literal: true
 
-require 'opentelemetry/sdk'
+require 'opentelemetry/sdk' if ENV['SIGNOZ_MONITORING_ENABLED'] == 'true'
 require_relative 'application'
 
-OpenTelemetry::SDK.configure do |c|
-  c.use_all
-end
+OpenTelemetry::SDK.configure(&:use_all) if ENV['SIGNOZ_MONITORING_ENABLED'] == 'true'
 
 Rails.application.initialize!