+++ spec/factories/stats.rb
@@ -1,8 +1,19 @@
+# frozen_string_literal: true
+
 FactoryBot.define do
   factory :stat do
     year { 1 }
     month { 1 }
     distance { 1 }
-    toponyms { "" }
+    toponyms do
+      [
+        {
+          'cities' => [
+            { 'city' => 'Moscow', 'points' => 7, 'timestamp' => 1_554_317_696, 'stayed_for' => 1831 }
+          ],
+          'country' => 'Russia'
+        }, { 'cities' => [], 'country' => nil }
+      ]
+    end
   end
 end