+++ spec/jobs/area_visits_calculating_job_spec.rb
@@ -1,5 +1,16 @@
+# frozen_string_literal: true
+
 require 'rails_helper'
 
 RSpec.describe AreaVisitsCalculatingJob, type: :job do
-  pending "add some examples to (or delete) #{__FILE__}"
+  describe '#perform' do
+    let(:user) { create(:user) }
+    let(:area) { create(:area, user:) }
+
+    it 'calls the AreaVisitsCalculationService' do
+      expect(Areas::Visits::Create).to receive(:new).with(user, [area]).and_call_original
+
+      described_class.new.perform(user.id)
+    end
+  end
 end