+++ CHANGELOG.md
@@ -5,6 +5,34 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+### Added
+
+- Added OpenTelemetry to the application to collect metrics and traces via [Signoz](https://signoz.io/docs/instrumentation/ruby-on-rails/)
+
+## [0.9.9] — 2024-07-30
+
+### Added
+
+- Pagination to exports page
+- Pagination to imports page
+- GET `/api/v1/points` endpoint to get all points for the user with swagger docs
+- DELETE `/api/v1/points/:id` endpoint to delete a single point for the user with swagger docs
+- DELETE `/api/v1/areas/:id` swagger docs
+- User can now change route opacity in settings
+- Points on the Points page can now be ordered by oldest or newest points
+- Visits on the Visits page can now be ordered by oldest or newest visits
+
+### Changed
+
+- Point deletion is now being done using an api key instead of CSRF token
+
+### Fixed
+
+- OpenStreetMap layer is now being selected by default in map controls
+
+---
 
 ## [0.9.8] — 2024-07-27